Break apart ClientFactory
