+++ extensions/src/AWSSDK.Extensions.NETCore.Setup/IClientFactory.cs
@@ -1,0 +1,16 @@
+using Amazon.Runtime;
+
+namespace AWSSDK.Extensions.NETCore.Setup
+{
+    /// <summary>
+    ///
+    /// </summary>
+    public interface IClientFactory
+    {
+        /// <summary>
+        ///
+        /// </summary>
+        /// <returns></returns>
+        IAmazonService CreateServiceClient<T>(AWSCredentials credentials = null) where T : IAmazonService;
+    }
+}
