+++ extensions/src/AWSSDK.Extensions.NETCore.Setup/Impl/DefaultClientConfigFactory.cs
@@ -1,154 +1,28 @@
-﻿/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
- * 
- *  http://aws.amazon.com/apache2.0
- * 
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Reflection;
-using System.Xml.Linq;
+using Amazon.Extensions.NETCore.Setup;
 using Amazon.Runtime;
-using Amazon.Runtime.CredentialManagement;
-using Amazon.Runtime.Credentials.Internal;
-using AWSSDK.Extensions.NETCore.Setup;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 
-namespace Amazon.Extensions.NETCore.Setup
+namespace AWSSDK.Extensions.NETCore.Setup.Impl
 {
     /// <summary>
-    /// The factory class for creating AWS service clients from the AWS SDK for .NET.
+    ///
     /// </summary>
-    internal class ClientFactory<T>
-        where T : IAmazonService
+    public class DefaultClientConfigFactory : IClientConfigFactory
     {
         private static readonly Type[] EMPTY_TYPES = Array.Empty<Type>();
         private static readonly object[] EMPTY_PARAMETERS = Array.Empty<object>();
-
-        private readonly AWSOptions _options;
-        private readonly IAWSCredentialsFactory _credentialsFactory;
-        private readonly ILogger _logger;
-
-        /// <summary>
-        /// Constructs an instance of the ClientFactory
-        /// </summary>
-        /// <param name="awsOptions">The AWS options used for creating service clients.</param>
-        /// <param name="credentialsFactory"></param>
-        /// <param name="logger"></param>
-        internal ClientFactory(AWSOptions awsOptions, IAWSCredentialsFactory credentialsFactory, ILogger logger)
-        {
-            _options = awsOptions ?? throw new ArgumentNullException(nameof(awsOptions));
-            _credentialsFactory = credentialsFactory ?? throw new ArgumentNullException(nameof(credentialsFactory));
-            _logger = logger;
-        }
-
-        /// <summary>
-        /// Creates the AWS service client that implements the service client interface.
-        /// </summary>
-        /// <returns>The AWS service client</returns>
-        internal IAmazonService CreateServiceClient(AWSCredentials credentials = null)
-        {
-            PerformGlobalConfig(_logger, _options);
-            credentials = credentials ?? _credentialsFactory.Create();
-
-            if (!string.IsNullOrEmpty(_options?.SessionRoleArn))
-            {
-                if (string.IsNullOrEmpty(_options?.ExternalId))
-                {
-                    credentials = new AssumeRoleAWSCredentials(credentials, _options.SessionRoleArn, _options.SessionName);
-                }
-                else
-                {
-                    credentials = new AssumeRoleAWSCredentials(credentials, _options.SessionRoleArn, _options.SessionName, new AssumeRoleAWSCredentialsOptions() { ExternalId = _options.ExternalId });
-                }
-            }
-
-            var config = CreateConfig(_options);
-            var client = CreateClient(credentials, config);
-            return client as IAmazonService;
-        }
-
-        /// <summary>
-        /// Performs all of the global settings that have been specified in AWSOptions.
-        /// </summary>
-        /// <param name="logger"></param>
-        /// <param name="options"></param>
-        private static void PerformGlobalConfig(ILogger logger, AWSOptions options)
-        {
-            if(options?.Logging != null)
-            {
-                if(options.Logging.LogTo.HasValue && AWSConfigs.LoggingConfig.LogTo != options.Logging.LogTo.Value)
-                {
-                    AWSConfigs.LoggingConfig.LogTo = options.Logging.LogTo.Value;
-                    logger?.LogDebug($"Configuring SDK LogTo: {AWSConfigs.LoggingConfig.LogTo}");
-                }
-
-                if (options.Logging.LogResponses.HasValue && AWSConfigs.LoggingConfig.LogResponses != options.Logging.LogResponses.Value)
-                {
-                    AWSConfigs.LoggingConfig.LogResponses = options.Logging.LogResponses.Value;
-                    logger?.LogDebug($"Configuring SDK LogResponses: {AWSConfigs.LoggingConfig.LogResponses}");
-                }
-
-                if (options.Logging.LogMetrics.HasValue && AWSConfigs.LoggingConfig.LogMetrics != options.Logging.LogMetrics.Value)
-                {
-                    AWSConfigs.LoggingConfig.LogMetrics = options.Logging.LogMetrics.Value;
-                    logger?.LogDebug($"Configuring SDK LogMetrics: {AWSConfigs.LoggingConfig.LogMetrics}");
-                }
-
-                if (options.Logging.LogResponsesSizeLimit.HasValue && AWSConfigs.LoggingConfig.LogResponsesSizeLimit != options.Logging.LogResponsesSizeLimit.Value)
-                {
-                    AWSConfigs.LoggingConfig.LogResponsesSizeLimit = options.Logging.LogResponsesSizeLimit.Value;
-                    logger?.LogDebug($"Configuring SDK LogResponsesSizeLimit: {AWSConfigs.LoggingConfig.LogResponsesSizeLimit}");
-                }
-            }
-        }
-
-        /// <summary>
-        /// Creates the service client using the credentials and client config.
-        /// </summary>
-        /// <param name="credentials"></param>
-        /// <param name="config"></param>
-        /// <returns></returns>
-        private static AmazonServiceClient CreateClient(AWSCredentials credentials, ClientConfig config)
-        {
-#if NET8_0_OR_GREATER
-            return T.CreateDefaultServiceClient(credentials, config) as AmazonServiceClient;
-#else
-            var clientTypeName = typeof(T).Namespace + "." + typeof(T).Name.Substring(1) + "Client";
-            var clientType = typeof(T).GetTypeInfo().Assembly.GetType(clientTypeName);
-            if (clientType == null)
-            {
-                throw new AmazonClientException($"Failed to find service client {clientTypeName} which implements {typeof(T).FullName}.");
-            }
-
-            var constructor = clientType.GetConstructor(new Type[] { typeof(AWSCredentials), config.GetType() });
-            if (constructor == null)
-            {
-                throw new AmazonClientException($"Service client {clientTypeName} missing a constructor with parameters AWSCredentials and {config.GetType().FullName}.");
-            }
-
-            return constructor.Invoke(new object[] { credentials, config }) as AmazonServiceClient;
-#endif
-        }
 
         /// <summary>
         /// Creates the ClientConfig object for the service client.
         /// </summary>
         /// <param name="options"></param>
         /// <returns></returns>
-        private ClientConfig CreateConfig(AWSOptions options)
+        public ClientConfig CreateConfig<T>(AWSOptions options)
+            where T : IAmazonService
         {
 #if NET8_0_OR_GREATER
             ClientConfig config = T.CreateDefaultClientConfig();
